+++ Mahmoud/Global-State-Management/react-course-part2-starter-main/src/App.tsx
@@ -2,7 +2,7 @@ import "./App.css";
 import AuthProvider from "./state-management/AuthProvider";
 import HomePage from "./state-management/HomePage";
 import NavBar from "./state-management/NavBar";
-import TaskProvider from "./state-management/TaskProvider";
+import TaskProvider from "./state-management/tasks/TaskProvider";
 
 function App() {
   return (
