--- Mahmoud/Global-State-Management/react-course-part2-starter-main/src/state-management/TaskProvider.tsx
@@ -1,19 +1,0 @@
-import { ReactNode, useReducer } from "react";
-import TaskContext from "./contexts/taskContext";
-import tasksReducer from "./reducers/taskReducer";
-
-interface Props {
-  children: ReactNode;
-}
-
-const TaskProvider = ({ children }: Props) => {
-  const [tasks, dispatch] = useReducer(tasksReducer, []);
-
-  return (
-    <TaskContext.Provider value={{ tasks, dispatch }}>
-      {children}
-    </TaskContext.Provider>
-  );
-};
-
-export default TaskProvider;
