--- Mahmoud/Global-State-Management/react-course-part2-starter-main/src/state-management/hooks/useTasks.ts
@@ -1,6 +1,0 @@
-import { useContext } from "react";
-import TaskContext from "../contexts/taskContext";
-
-const useTasks = () => useContext(TaskContext);
-
-export default useTasks;
