+++ Mahmoud/Global-State-Management/react-course-part2-starter-main/src/state-management/tasks/TaskProvider.tsx
@@ -1,3 +1,6 @@
+import { ReactNode, useReducer } from "react";
+import TaskContext from "./taskContext";
+
 export interface Task {
   id: number;
   title: string;
@@ -24,4 +27,18 @@ const tasksReducer = (tasks: Task[], action: TaskAction): Task[] => {
   }
 };
 
-export default tasksReducer;
+interface Props {
+  children: ReactNode;
+}
+
+const TaskProvider = ({ children }: Props) => {
+  const [tasks, dispatch] = useReducer(tasksReducer, []);
+
+  return (
+    <TaskContext.Provider value={{ tasks, dispatch }}>
+      {children}
+    </TaskContext.Provider>
+  );
+};
+
+export default TaskProvider;
