+++ Mahmoud/Global-State-Management/react-course-part2-starter-main/src/state-management/tasks/index.ts
@@ -1,0 +1,2 @@
+export { default as TaskProvider } from "./TaskProvider";
+export { default as TaskContext } from "./TaskList";
