+++ Mahmoud/Global-State-Management/react-course-part2-starter-main/src/state-management/tasks/taskContext.ts
@@ -1,6 +1,6 @@
 import { Dispatch } from "react";
-import { Task, TaskAction } from "../reducers/taskReducer";
 import React from "react";
+import { Task, TaskAction } from "./TaskProvider";
 
 interface TaskContextType {
   tasks: Task[];
